langgraph-rag/requirements.txt
python-dotenv>=1.0.1
langgraph>=0.2.35
langchain>=0.3.4
faiss-cpu>=1.8.0
numpy>=1.26,<2.0
scikit-learn>=1.4
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
tiktoken>=0.7
pydantic>=2.8
streamlit>=1.38
rich>=13.7
typer>=0.12
pytest>=8.2